backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.1.post1
insightface==0.7.3
onnxruntime==1.17.0
opencv-python-headless==4.9.0.80
python-dotenv==1.0.1
cloudinary==1.40.0
